framework/config: panic on invalid config file
When a config yaml can not be parsed and config debug-log is disabled,
you will not notice that the config is not loaded.

Since an existing config file can always be assumed as "wanted", we
should panic if it can not be parsed instead of starting gracefully
without the config. The same applies for additional config provided via
flamingo-config flag.