framework: Fix graceful shutdown that was blocked by prefix router, avoid duplicated log's (#177)
The prefix router module was listening to the wrong event and therefore was never shut down. This resulted in a hanging state after pressing CTRL+C.

Since there are two ways of shutting down a command (signal for infinite command, post run for finite command) using the signal one lead to duplicate call of the shutdown function / log out put. This is now wrapped in a sync.Once to only trigger the shutdown once.

In addition the default serve command tried to shutdown it's server independed of if it was running, this is now also checked to avoid confusing log outputs.