framework/prefixrouter: Fix potential nil pointer when using malformed requests
The front router truncates the prefix in the path and creates a new request. To do this the url.Parse function is used. If an invalid path is processed a nil pointer occurs.

This PR changes the behavior and returns a 404 page if an invalid path is used in the request.

fixes #198