feat(core/auth/oauth): Add optional OIDC callback error handler
In some cases you want to handle errors that happen during the OIDC callback differently.

A use case could be authenticating the user using the `prompt=none` flow which can result in a `error=login_required` on the Callback endpoint if the user is not already signed in with the identity provider. In this case you might want to redirect the user or display a specific error page.

The handler is optional and if it can't handle the error the build in logic kicks in.