feat(core/auth/oidc): Add support of `AuthCodeOption` in the token exchange during callback (#338)
* feat(core/auth/oidc): Add support of `AuthCodeOption` in the token exchange during callback

Currently, it is only possible to modify the authorization URL using the `AuthCodeOptioner` interface. Since the Exchange function also supports this pattern we should also pass the registered `AuthCodeOptioner` to it.

For example, both my authorization and my token retrieval endpoint need some custom URL param to work. Currently, only the one for the auth endpoint would be possible.

---------

Co-authored-by: Karol Nowak <karol.nowak@omnevo.net>